Move app handler resource to admin::rest_api::actix::ws_register_type
